% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{a.b.main.image.recognition}
\alias{a.b.main.image.recognition}
\title{MAIN - Image Recognition}
\usage{
a.b.main.image.recognition()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

Meta function for starting meta steps of the analysis.

Each step will be further explained in the corresponding function descriptions
}
\author{
Vitali Friesen
}
