% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.a.preprocessing.start}
\alias{b.a.preprocessing.start}
\title{Feature Extraction -  Wrapper function}
\usage{
b.a.preprocessing.start()
}
\description{
The needed package /code{EBImage} can be installed by using the source /code{\linkg{"https://bioconductor.org/biocLite.R"}} then it can be loaded
with the command /code{biocLite("EBImage")}. To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for ...

It executes these functions:
\enumerate{
  \item \code{\link{b.b.createTrainTestDataIndexes}}
  \item \code{\link{b.c.loadClassData}}
}
}
\author{
Vitali Friesen, Tassilo Tobollik
}
