% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.b.createTrainTestDataIndexes}
\alias{b.b.createTrainTestDataIndexes}
\title{Preprocessing - Create Train Test DataSet Indexes}
\usage{
b.b.createTrainTestDataIndexes()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

...
}
\author{
Vitali Friesen, Tassilo Tobollik
}
