% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.c.loadClassData}
\alias{b.c.loadClassData}
\title{Feature Extraction - Step 2}
\usage{
b.c.loadClassData()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

...
}
\author{
Vitali Friesen
}
