% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.d.normalization.start}
\alias{b.d.normalization.start}
\title{Preprocessing - Method that starts either the rgb-normalization or the histogram-equalization of an image in this project}
\usage{
b.d.normalization.start(rgbNorm = T)
}
\arguments{
\item{rgbNorm}{A boolean that describes whether RGB-normalization (/code{TRUE}) or histogram-equalization (/code{FALSE}) should be used}
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Tassilo Tobollik
}
