% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.d.rgbNorm}
\alias{b.d.rgbNorm}
\title{Preprocessing - Normalize rgb images to remove shadow and light effects}
\usage{
b.d.rgbNorm(filename, x = 480, y = 640, z = 3)
}
\arguments{
\item{filename}{a string that describes the path to an image from the workingdirectory and it's filename}

\item{x}{A number that describes the images horizontal number of pixels}

\item{y}{A number that describes the images vertical number of pixeks}

\item{z}{A number that describes the images number of color dimensions (normally 3 for RGB or 1 for grayscale)}
}
\value{
\code{normImg} an array of size \code{x * y} and with \code{z} dimensions that holds the normalized RGB-values of the image described by \code{filename}
}
\description{
The value of each pixel, of the inserted image, in each colour-dimension will be divided by the sum of the pixels values in all dimensions.
By this shadows and light highlights in the image are removed to some degree.
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\examples{
\\code{b.d.rgbNorm(filename = "data/image.png", x = 480, y = 640, z = 3)}
}
\author{
Tassilo Tobollik
}
