% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.e.transformGrayscale}
\alias{b.e.transformGrayscale}
\title{Preprocessing - Transform rgb images to grayscale}
\usage{
b.e.transformGrayscale(image)
}
\arguments{
\item{image}{An array with three dimensions that holds the values of an RGB-image}
}
\value{
An array with one dimension that holds the values of a grayscale version of the inserted image
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Tassilo Tobollik
}
