% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.f.transformHistEqualRgb}
\alias{b.f.transformHistEqualRgb}
\title{Preprocessing - Transform rgb images using histogram equalization}
\usage{
b.f.transformHistEqualRgb(image = NULL, filename)
}
\arguments{
\item{image}{An array that can have multiple dimensions and holds the values of image (for example an RGB image). This parameter is optional.}

\item{filename}{A string that holds the path to the image from the workingdirectory and it's filename. If the parameter \code{image} isn't given the image will be loaded 
by using this path.}
}
\value{
An image with histogram equalized RGB-values
}
\description{
The needed package /code{EBImage} can be installed by using the source /code{\linkg{"https://bioconductor.org/biocLite.R"}} then it can be loaded
with the command /code{biocLite("EBImage")}. To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\examples{
\\code{histEqualImage <- b.f.transformHistEqualRgb(imageArray, "data/image.png")}
}
\author{
Tassilo Tobollik
Takes aprox. 0.38 sec/image
}
