% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.g.transformHistEqualGray}
\alias{b.g.transformHistEqualGray}
\title{Preprocessing - Transform grayscale images using histogram equalization}
\usage{
b.g.transformHistEqualGray(image = NULL, filename)
}
\arguments{
\item{image}{An array with one dimensions that holds the values of a grayscale image. This parameter is optional.}

\item{filename}{A string that holds the path to the image from the workingdirectory and it's filename. If the parameter \code{image} isn't given the image will be loaded 
by using this path.}
}
\value{
An image with histogram equalized grayscale values
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\examples{
\\code{histEqualImage <- b.f.transformHistEqualGray(imageArray, "data/image.png")}
}
\author{
Tassilo Tobollik
}
