% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.h.displayImageHist}
\alias{b.h.displayImageHist}
\title{Preprocessing - Display the histogram of an image}
\usage{
b.h.displayImageHist(image)
}
\arguments{
\item{image}{An array with three dimensions that holds the values of an RGB-image}
}
\value{
A chart that displays the histogram of the inserted image
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Tassilo Tobollik
}
