% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.j.assignSimpelAccuracy}
\alias{b.j.assignSimpelAccuracy}
\title{Preprocessing -}
\usage{
b.j.assignSimpelAccuracy(image)
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

...
}
\author{
Vitali Friesen
}
