% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{b.k.rgbNorm}
\alias{b.k.rgbNorm}
\title{Preprocessing -}
\usage{
b.k.rgbNorm(filename, x = 480, y = 640, z = 3)
}
\arguments{
\item{filename}{A string that holds the filename}

\item{x}{A number that describes the height of the image in pixels}

\item{y}{A number that describes the width of the image in pixels}

\item{z}{A number that describes the number of colour dimensions of the image}
}
\description{
Set \code{setwd()} to the location of the image.
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Tassilo Tobollik
}
