% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{c.a.feature.start}
\alias{c.a.feature.start}
\title{Feature Extraction -  Wrapper function}
\usage{
\method{c}{a.feature.start}()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for ...

It executes these functions:
\enumerate{
  \item \code{\link{c.b.colorHist}}
  \item \code{\link{c.c.colorHistPlotting}}
}
}
\author{
Vitali Friesen
}
