% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{c.b.colorHist}
\alias{c.b.colorHist}
\title{Feature Extraction - Step 1}
\usage{
\method{c}{b.colorHist}()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

...
}
\author{
Vitali Friesen
}
