% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{c.c.colorHistPlotting}
\alias{c.c.colorHistPlotting}
\title{Feature Extraction - Plotting Histograms}
\usage{
\method{c}{c.colorHistPlotting}(picture = "hl 2017-06-14 17-44-58-48.png")
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

...
}
\author{
Vitali Friesen
}
