% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{c.d.hog}
\alias{c.d.hog}
\title{Feature Extraction - Step 2 (Histogram of Oriented Gradients)}
\usage{
\method{c}{d.hog}(cells, orientations)
}
\arguments{
\item{cells}{A number that describes the hog folding and by that into how many \code{cells*cells} the image will be cutted}

\item{orientations}{A number that describes how many orientation vectors are stored for each hog cell. This number is for the vectors from 1?-180?}
}
\value{
A dataframe that holds the hog feature data for the images in the folder \code{"data-raw/IMG/CS CZ/normal"} of the package
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
Set working directory to the package before using this function (setwd())
}
\author{
Sascha Di Bernardo, Tassilo Tobollik
}
