% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{c.f.plotHogCellVectors}
\alias{c.f.plotHogCellVectors}
\title{Feature Extraction - plot HOG (Histogram of Oriented Gradients) cell vectors}
\usage{
\method{c}{f.plotHogCellVectors}(vectorValues, cellCol, cellRow)
}
\arguments{
\item{vectorValues}{A vector which describes the lengths of the vectors in the cell}

\item{cellCol}{A number that describes the column of the rastered image in which the cell lies}

\item{cellRow}{A number that describes the row of the rastered image in which the cell lies}
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
 Function that plots vectors with given angles and length into a specified layout cell
}
\author{
Tassilo Tobollik
}
