% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{c.g.getPixelInfo}
\alias{c.g.getPixelInfo}
\title{Feature Extraction -}
\usage{
\method{c}{g.getPixelInfo}()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
 
...
}
\author{
Vitali Friesen
}
