% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{d.a.randomForest.start}
\alias{d.a.randomForest.start}
\title{Random Forest - Wrapper function}
\usage{
d.a.randomForest.start(block, feature1, saveFile, classLabel, numTrees,
  feature2 = NULL, mtry = NULL, nodesize = NULL, cutoff = NULL)
}
\arguments{
\item{block}{A string that holds a path to a cross-validation block file}

\item{feature1}{A string that holds a path to a feature data file (hog or colorHist)}

\item{saveFile}{A string that holds a path to which the result should be stored as .rda file.}

\item{classLabel}{A string that holds a path to a file with classification labels for the feature data}

\item{numTrees}{A number that defines the number of trees randomForest should use (see package randomForest)}

\item{feature2}{A string that holds a path to a second feature data file (the type that was not used in \code{feature1}).This parameter is optional}

\item{mtry}{A number that defines the number of variables randomly sampled as candidates at each split of randomForest (see package randomForest). This parameter is optional}

\item{nodesize}{A number that defines the minimum size of terminal nodes for randomForest (see package randomForest). This parameter is optional}

\item{cutoff}{A vector of length equal to number of classes that defines the ratio of proportion of votes to cutoff (see package randomForest). This parameter is optional}
}
\value{
A table that holds the pairs of correct and not correct predicted images, the number of correct predicted images and the accuracy percentage
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for a random forest classifier.

It executes these functions:
\enumerate{
  \item \code{\link{d.b.step1}}
  \item \code{\link{d.c.step2}}
  \item \code{\link{d.d.evaluation}}
  \item \code{\link{d.e.plotImportanceColorHist}}
}
}
\examples{
\\code{result <- d.a.randomForest.start("data/blocks2677IMG.rda", "data/hog_original_8_9_complete.rda",
"data/colorHistRGBNorm_original_255buckets_hog_original_8_9_nodes7_rf100_result.rda", "data/classesOrig.rda",
100, feature2 = "data/colorHistOriginalRGBNorm255Buckets.rda",nodesize = 7)}
}
\author{
Vitali Friesen, Colin Juers, Tassilo Tobollik
}
