% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{d.b.step1}
\alias{d.b.step1}
\title{Classifier 1 - Step 1}
\usage{
d.b.step1(trainData, numTrees, pMtry = NULL, pNodesize = NULL,
  pCutoff = NULL)
}
\arguments{
\item{trainData}{A data frame with number of images as rowsize and number of features as columnsize that holds the feature data}

\item{numTrees}{A number that defines the number of trees randomForest should use (see package randomForest)}

\item{pMtry}{A number that defines the number of variables randomly sampled as candidates at each split of randomForest (see package randomForest). This parameter is optional}

\item{pNodesize}{A number that defines the minimum size of terminal nodes for randomForest (see package randomForest). This parameter is optional}

\item{pCutoff}{A vector of length equal to number of classes that defines the ratio of proportion of votes to cutoff (see package randomForest). This parameter is optional}
}
\value{
A randomForest model that has been trained on the feature data
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\examples{
\\code{rfModel <- d.b.step1(featureData, numTrees, mtry, nodesize, cutoff)}
}
\author{
Vitali Friesen, Colin Juers, Tassilo Tobollik
}
