% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{d.c.step2}
\alias{d.c.step2}
\title{Random Forest - Step 2}
\usage{
d.c.step2(testData, rfModel)
}
\arguments{
\item{testData}{A dataframe with number of images as rowsize and number of features as columnsize that holds the feature data which should be predicted}

\item{rfModel}{A randomForest classification model}
}
\value{
A vector that holds the prediction for \code{testData} with length equal to the number of images in the dataset
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\examples{
\\code{pred <- d.c.step2(testData, rfModel)}
}
\author{
Vitali Friesen, Colin Juers, Tassilo Tobollik
}
