% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{d.d.evaluation}
\alias{d.d.evaluation}
\title{Random Forrest - Evaluation}
\usage{
d.d.evaluation(pred, testData)
}
\arguments{
\item{pred}{A vector with all predictions for \code{P} of all iterations}

\item{testData}{A Vector with all actual values for \code{P} of all iterations}
}
\value{
A matrix with the correct predictions, the error predictions and the accuracy
}
\description{
Evaluates the predicted values for \code{P} of the testset with the actual values. Therefore, creates
table with right and wrong preditions and calculates the error 1. and 2. degree. Also, gives visual evaluation with bar chart
and pie chart. 
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Colin Juers, Tassilo Tobollik
}
