% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{d.e.plotImportanceColorHist}
\alias{d.e.plotImportanceColorHist}
\title{Classifier 1 - Step 1}
\usage{
d.e.plotImportanceColorHist(rfModels)
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

...
}
\author{
Vitali Friesen, Colin Juers, Tassilo Tobollik
}
