% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{e.a.ann.start}
\alias{e.a.ann.start}
\title{Artifical Neural Network - Wrapper function}
\usage{
e.a.ann.start(blocks, feature1, feature2 = NULL, classification, dataType,
  imageType, rounds, lr, nodes, nodeString, batch)
}
\arguments{
\item{blocks}{A string that holds a path to a cross-validation block file}

\item{feature1}{A string that holds a path to a feature data file (hog, colorHist or pixelFeature28)}

\item{feature2}{A string that holds a path to a second feature data file (the type that was not used in \code{feature1}).This parameter is optional}

\item{classification}{A string that holds a path to which the result should be stored as .rda file.}

\item{dataType}{er that defines the number of trees randomForest should use (see package randomForest)}

\item{imageType}{A number that defines the number of variables randomly sampled as candidates at each split of randomForest (see package randomForest). This parameter is optional}

\item{rounds}{A number that defines the minimum size of terminal nodes for randomForest (see package randomForest). This parameter is optional}

\item{lr}{A vector of length equal to number of classes that defines the ratio of proportion of votes to cutoff (see package randomForest). This parameter is optional}

\item{nodes}{A vector containing node amounts per layer, e.g. rep(1000, 2) for two layers with 1000 nodes.}

\item{nodeString}{String representation of nodes that are used}

\item{batch}{The amount of objects which the model is trained on every round}
}
\value{
A table that holds the pairs of correct and not correct predicted images, the number of correct predicted images and the accuracy percentage
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for an artifical neural network

It executes these functions:
\enumerate{
  \item \code{\link{e.b.ann.step1}}
}
}
\examples{
\\code{result <- e.a.ann.start(blocks = "data/blocks2677IMG.rda", feature1 = "data/hog_original_15_6_complete.rda", classification = "data/classesOrig.rda", 
  dataType = "hog_15_6", imageType = "orig", rounds = "50", lr = "0.00001",
  nodes = rep(10,100), nodeString = "10x100", batch = "100")}
}
\author{
Nils Meckmann, Maren Reuter, Sascha di Bernardo
}
