% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{f.a.cnn.start}
\alias{f.a.cnn.start}
\title{Convolutional Neural Network - Wrapper function}
\usage{
f.a.cnn.start(rounds, lr, batch)
}
\arguments{
\item{rounds}{A numeric value that is the amount of rounds the model is trained}

\item{lr}{The learning rate}

\item{batch}{The amount of objects which the model is trained on every round}
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for a convolutional neural network classifier.

It executes these functions:
\enumerate{
  \item \code{\link{f.b.cnn.step1}}
  \item \code{\link{f.b.cnn.step2}}
}
}
\examples{
\\code{result -> f.a.cnn.start(rounds = 80, lr = 0.00000001, batch = 50)}

}
\author{
Maren Reuter, Nils Meckmann
}
