% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{f.c.cnn.step2}
\alias{f.c.cnn.step2}
\title{Convolutional Neural Network - Step 2}
\usage{
f.c.cnn.step2(CNN_model, test_array)
}
\arguments{
\item{test_array}{An data array with features to test the CNN_Model, rows being objects and columns being features}

\item{CNN_Model}{the trained CNN_Model}
}
\value{
preds predictions of the CNN_Modelregarding the test data set
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Maren Reuter, Nils Meckmann
}
