% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{g.a.svm.start}
\alias{g.a.svm.start}
\title{Classifier Support Vector Machine -  Wrapper function}
\usage{
g.a.svm.start(block, feature1, saveFile, classLabel, feature2 = NULL)
}
\arguments{
\item{block}{A string that holds a path to a cross-validation block file}

\item{feature1}{A string that holds a path to a feature data file (hog or colorHist)}

\item{saveFile}{A string that holds a path to which the result should be stored as .rda file.}

\item{classLabel}{A string that holds a path to a file with classification labels for the feature data}

\item{feature2}{A string that holds a path to a second feature data file (the type that was not used in \code{feature1}).This parameter is optional}
}
\value{
A table that holds the pairs of correct and not correct predicted images, the number of correct predicted images and the accuracy percentage
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for classifying data with support vector machines

It executes these functions:
\enumerate{
  \item \code{\link{g.b.step1}}
  \item \code{\link{g.c.step2}}
}
}
\examples{
\\code{result <- d.a.randomForest.start("data/blocks2677IMG.rda", "data/hog_original_8_9_complete.rda",
"data/colorHistRGBNorm_original_255buckets_hog_original_8_9_nodes7_svm_result.rda", "data/classesOrig.rda",
100, feature2 = "data/colorHistOriginalRGBNorm255Buckets.rda")}
}
\author{
Colin Juers
}
