% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{g.b.step1}
\alias{g.b.step1}
\title{Classifier Support Vector Machine - Step 1}
\usage{
g.b.step1(trainData)
}
\arguments{
\item{trainData}{A sample matrix of all files to train the svm classifier}
}
\value{
Support Vector Machine model trained with sampled data
}
\description{
The train matrix is used to compute hyperplanes of the feature vector. The parameters of the function have
been tuned with the \code{tune()} function.
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Colin Juers
}
