% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{g.c.step2}
\alias{g.c.step2}
\title{Classifier Support Vector Machine - Step 2}
\usage{
g.c.step2(model_svm, testData)
}
\arguments{
\item{model_svm}{The trained svm model with all hyperplanes}

\item{testData}{A sample matrix of all files to apply the trained svm model on and predict "P"}
}
\value{
Vector of predicted lables for the test set
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}
}
\author{
Colin Juers
}
