% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.a.util.gen.summary}
\alias{z.a.util.gen.summary}
\title{Util - Generate Summary}
\usage{
z.a.util.gen.summary(frame1, f1_type, frame2 = NULL, ...)
}
\arguments{
\item{frame1}{A data.frame or matrix which get summarzed.}

\item{f1_type}{A name of type character for the context the observations in the data represent. These could by "Spam" for example. It then gets prefixed to the element of summarization of the row like "Spam Mean:".

I will only be used if one data table is provided. For two data table it will be "Diff" by standard and will create row names like "Diff Mean:".}

\item{frame2}{A data.frame or matrix which get summarzed and compared to frame1. Number and name of the columns of \code{frame1} and \code{frame2} need to be identical}
}
\value{
A data.frame with either 6 rows for ein provided data.frame with the elements min, 1st quantile, median, mean, 3rd quantile and max.

For two data.frames it contains 20 rows with 6 rows looking explained above, another 6 rows showing the difference of each summary element of the two data tables and 2 rows of empty space between each summary.
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}

Generate a summary of given data or compare two

This function generates a summary of a data.frame with one or two given tables with data.
Each column/attribute of the given data will be summarized individually.

It basically works like the standard summary function of the \code{base} package but it does not store the row description in each individual cell like it is done in the standard function.

Furthermore it also compares the summary of to data tables if two are passed as parameters.
}
\author{
Vitali Friesen
}
