% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.b.util.diff}
\alias{z.b.util.diff}
\title{Util - DIfference}
\usage{
z.b.util.diff(vec1, vec2)
}
\arguments{
\item{vec1}{A numeric vector.}

\item{vec2}{A numeric vector of the same size as \code{vec1}}
}
\value{
A character vector containing the relative sizes. Special cases for \code{0} elements in the vector.
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}

Support function for \code{\link{z.a.util.gen.summary}}

Takes two numeric vectors and calculates the relative size of the numbers at the same positions.
}
\examples{
z.b.util.diff(c(8,0,10,0),c(4,3,15,0))
# [1] "50\%" "only in 2nd" "150\%" "Both 0"

}
\author{
Vitali Friesen
}
