% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.e.RGBAtoRGB}
\alias{z.e.RGBAtoRGB}
\title{Util - Function that converts RGBA to RGB}
\usage{
z.e.RGBAtoRGB(r = 0, g = 0, b = 0, a, r2, g2, b2)
}
\arguments{
\item{r}{An one dimensional dataframe that holds the red colour values of the background of the image}

\item{g}{An one dimensional dataframe that holds the green colour values of the background of the image}

\item{b}{An one dimensional dataframe that holds the blue colour values of the background of the image}

\item{a}{An one dimensional dataframe that holds the alpha values of the image}

\item{r2}{An one dimensional dataframe that holds the red colour values of the foreground of the image}

\item{g2}{An one dimensional dataframe that holds the green colour values of the foreground of the image}

\item{b2}{An one dimensional dataframe that holds the blue colour values of the foreground of the image}
}
\value{
A three dimensional dataframe that holds the converted RGB colour values of the image
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}

This function transforms the given RGBA values to RGB values.
Thereby r,g,b are the background RGB colors, r2,g2,b2 are the foreground picture colors of the RGBA image
and a is the alpha value of the RGBA image.
}
\author{
Tassilo Tobollik
}
