% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.f.displayRgbImage}
\alias{z.f.displayRgbImage}
\title{Util - function that rasters and displays an image}
\usage{
z.f.displayRgbImage(image, x1 = NULL, x2 = NULL, y1 = NULL, y2 = NULL,
  interpol = F, noBorder = T)
}
\arguments{
\item{image}{A dataframe that holds the pixel values of the image that should be rendered}

\item{x1}{An optional number that describes the first x value of the part of the image that should be rendered}

\item{x2}{An optional number that describes the second x value of the part of the image that should be rendered}

\item{y1}{An optional number that describes the first y value of the part of the image that should be rendered}

\item{y2}{An optional number that describes the second y value of the part of the image that should be rendered}

\item{interpol}{A logical vector (or scalar) indicating whether to apply linear interpolation to the image when drawing}

\item{noBorder}{A boolean that describes if the image should be rendered without plot-borders}
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen, Tassilo Tobollik
}
