% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.g.readImage}
\alias{z.g.readImage}
\title{Util - function that reads in an image from the given path}
\usage{
z.g.readImage(path, asPng = F)
}
\arguments{
\item{path}{A string that holds a path to an image that should be read}

\item{asPng}{A boolean that describes if the image is an .png file or another image data type file}
}
\value{
A dataframe that holds the pixel values of the image
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Tassilo Tobollik
}
