% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.z.util.set.environment}
\alias{z.z.util.set.environment}
\title{Util - Environment Setting}
\usage{
z.z.util.set.environment()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}

This function simply sets some environment variables so they can be used across the package.
}
\author{
Vitali Friesen
}
